Implement shipment worker assignment and batch creation features; update frontend components for shipment management, including new buttons and modal interactions. Enhance backend services and repositories to support new functionalities, including worker assignment and batch shipment item creation.

+++ Full-Backend/backend/src/main/java/com/visera/backend/Repository/ShipmentItemRepository.java
@@ -5,9 +5,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface ShipmentItemRepository extends JpaRepository<ShipmentItem, Integer> {
+public interface ShipmentItemRepository extends JpaRepository<ShipmentItem, Long> {
 
-    List<ShipmentItem> findByShipmentId(int shipmentId);
+    List<ShipmentItem> findByShipmentId(Long shipmentId);
 
     List<ShipmentItem> findByStatus(String status);
 }
