+++ Full-Backend/backend/src/main/java/com/visera/backend/Repository/TaskRepository.java
@@ -5,12 +5,12 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface TaskRepository extends JpaRepository<Task, Integer> {
+public interface TaskRepository extends JpaRepository<Task, Long> {
 
-    List<Task> findByUserId(int userId);
+    List<Task> findByUserId(Long userId);
 
     List<Task> findByStatus(String status);
 
-    List<Task> findByShipmentItemId(int shipmentItemId);
+    List<Task> findByShipmentItemId(Long shipmentItemId);
 }
